Fix payments not working for business plan and move things around (#715)
* Add shared class that applies both business and ecom plan, move some things to payments class

* Cleanup for now-redundant stuff that's shared

* Fix formatting
