+++ store-on-wpcom/class-wc-calypso-bridge.php
@@ -79,7 +79,6 @@ class WC_Calypso_Bridge {
 		include_once dirname( __FILE__ ) . '/inc/wc-calypso-bridge-paypal-defaults.php';
 		include_once dirname( __FILE__ ) . '/inc/wc-calypso-bridge-paypal-method-supports.php';
 		include_once dirname( __FILE__ ) . '/inc/wc-calypso-bridge-products.php';
-		include_once dirname( __FILE__ ) . '/inc/wc-calypso-bridge-nav-unification-fixes.php';
 	}
 
 	/**
